requirements.txt
requests
notion-client
github-heatmap
retrying
pendulum
python-dotenv
html2notion
markdownify
mistletoe
md2notion
